# src/Bff/Example.Bff/openapi/Example.Bff_public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Example.Bff | public",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/sleepdiary/entries/{date}": {
      "get": {
        "tags": [
          "SleepDiary – SleepEntries"
        ],
        "summary": "Returns the caller's diary entry for one day.",
        "description": "Required scope: `sleepdiary.entry.read`.",
        "operationId": "SleepDiarySleepEntries_Get",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "Entry date (the day the user woke up), `YYYY-MM-DD`.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The entry for the day.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepDiarySleepEntryDto"
                }
              }
            }
          },
          "400": {
            "description": "The date is not in the `YYYY-MM-DD` format.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid access token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the scope `sleepdiary.entry.read` (`auth.missing_scope`) or does not identify a user (`auth.unauthenticated`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The caller has no entry for this day (`sleepdiary.entry.not_found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "SleepDiary – SleepEntries"
        ],
        "summary": "Records a new diary entry for the given day.",
        "description": "Required scope: `sleepdiary.entry.write`. One entry per day: to change an existing entry use `PUT`. The date may be at most\r\none day after today in UTC (tolerance for time zones ahead of UTC). Time in bed and sleep time (time in bed minus the time to fall asleep)\r\nare computed by the service. Recording an entry publishes the integration event `SleepEntryRecordedV1` for other services.",
        "operationId": "SleepDiarySleepEntries_Record",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "Entry date (the day the user woke up), `YYYY-MM-DD`.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "description": "The night's data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SleepDiarySleepEntryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SleepDiarySleepEntryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SleepDiarySleepEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The entry was recorded; the body contains its identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepDiaryCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data. `validation.failed`: quality outside 1–5, awakenings outside 0–50, negative time to fall asleep, note longer than\r\n2000 characters. A malformed date or body is rejected by the framework with a validation problem with `code` `request.malformed` (ADR-0044).\r\nBusiness rules: `sleepdiary.entry.future_date` (date too far in the future),\r\n`sleepdiary.entry.wake_before_bed` (wake-up not after bedtime), `sleepdiary.entry.wake_date_mismatch` (wake-up day differs\r\nfrom the entry date), `sleepdiary.entry.too_long` (more than 24 hours in bed), `sleepdiary.entry.invalid_latency`\r\n(time to fall asleep longer than the time in bed).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid access token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the scope `sleepdiary.entry.write` (`auth.missing_scope`) or does not identify a user (`auth.unauthenticated`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The caller already has an entry for this day (`sleepdiary.entry.already_exists`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "SleepDiary – SleepEntries"
        ],
        "summary": "Replaces all data of the caller's existing diary entry for the given day.",
        "description": "Required scope: `sleepdiary.entry.write`. A full replacement: every field is overwritten and an omitted or blank note removes the\r\nexisting one. The entry date cannot be changed; delete the entry and record a new one instead. Computed values are recalculated.\r\nNo integration event is published. When the request is rejected the entry stays unchanged.",
        "operationId": "SleepDiarySleepEntries_Update",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "Entry date (the day the user woke up), `YYYY-MM-DD`.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "description": "The new data of the night.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SleepDiarySleepEntryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SleepDiarySleepEntryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SleepDiarySleepEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "The entry was updated."
          },
          "400": {
            "description": "Invalid data. `validation.failed`: quality outside 1–5, awakenings outside 0–50, negative time to fall asleep, note longer than\r\n2000 characters. A malformed date or body is rejected by the framework with a validation problem with `code` `request.malformed` (ADR-0044). Business rules: `sleepdiary.entry.wake_before_bed`,\r\n`sleepdiary.entry.wake_date_mismatch`, `sleepdiary.entry.too_long`, `sleepdiary.entry.invalid_latency`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid access token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the scope `sleepdiary.entry.write` (`auth.missing_scope`) or does not identify a user (`auth.unauthenticated`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The caller has no entry for this day (`sleepdiary.entry.not_found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The entry was changed by another request between reading and saving (`persistence.concurrency_conflict`);\r\nreload it and retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "SleepDiary – SleepEntries"
        ],
        "summary": "Permanently deletes the caller's diary entry for the given day.",
        "description": "Required scope: `sleepdiary.entry.write`. A new entry can be recorded for the same day afterwards. Other services are not notified.",
        "operationId": "SleepDiarySleepEntries_Delete",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "Entry date (the day the user woke up), `YYYY-MM-DD`.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The entry was deleted."
          },
          "400": {
            "description": "The date is not in the `YYYY-MM-DD` format.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid access token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the scope `sleepdiary.entry.write` (`auth.missing_scope`) or does not identify a user (`auth.unauthenticated`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The caller has no entry for this day (`sleepdiary.entry.not_found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The entry was changed by another request between reading and saving (`persistence.concurrency_conflict`);\r\nreload it and retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. knowledge.material.archived; clients branch on it (ADR-0015)."
          },
          "traceId": {
            "type": "string",
            "description": "Trace identifier of the request, used to find its traces and logs."
          }
        }
      },
...
      "SleepDiaryCreatedResponse": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Technical identifier of the new entry (GUID). The entry is addressed by its date in all other endpoints; the identifier is informational\r\n<br />and matches `entryId` of the `SleepEntryRecordedV1` integration event.",
            "format": "uuid"
          }
        },
        "description": "Response body of `201 Created` after a new sleep diary entry was recorded."
      },
...
      "SleepDiarySleepEntryDto": {
        "required": [
          "id",
          "date",
          "bedTime",
          "wakeTime"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Technical identifier of the entry (GUID). Entries are addressed by Date in the API.",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "description": "Entry date: the day the user woke up.",
            "format": "date"
          },
          "bedTime": {
            "type": "string",
            "description": "When the user went to bed, local time without a time zone (usually on the day before Date).",
            "format": "date-time"
          },
          "wakeTime": {
            "type": "string",
            "description": "When the user got up, local time without a time zone; always later than BedTime and on the day Date.",
            "format": "date-time"
          },
          "sleepLatencyMinutes": {
            "type": "integer",
            "description": "How long it took to fall asleep, in minutes; between 0 and TimeInBedMinutes.",
            "format": "int32"
          },
          "awakenings": {
            "type": "integer",
            "description": "Number of awakenings during the night, from 0 to 50.",
            "format": "int32"
          },
          "quality": {
            "type": "integer",
            "description": "Subjective sleep quality, from 1 (worst) to 5 (best).",
            "format": "int32"
          },
          "notes": {
            "type": "string",
            "description": "The user's note, trimmed; `null` when there is none."
          },
          "timeInBedMinutes": {
            "type": "integer",
            "description": "Derived time in bed in whole minutes, from BedTime to WakeTime; at most 1440.",
            "format": "int32"
          },
          "sleepMinutes": {
            "type": "integer",
            "description": "Derived sleep time in minutes: TimeInBedMinutes minus SleepLatencyMinutes.",
            "format": "int32"
          }
        },
        "description": "One sleep diary entry as returned by the API, by GetSleepEntry and inside `SleepDiaryDto`."
      },
      "SleepDiarySleepEntryRequest": {
        "required": [
          "bedTime",
          "wakeTime"
        ],
        "type": "object",
        "properties": {
          "bedTime": {
            "type": "string",
            "description": "When the user went to bed, local date and time without an offset, e.g. `2026-09-29T23:15:00`; a value with `Z` or an offset\r\n<br />is rejected (HTTP 400).",
            "format": "date-time"
          },
          "wakeTime": {
            "type": "string",
            "description": "When the user got up, local date and time without an offset (a value with `Z` or an offset is rejected, HTTP 400). Must be later than BedTime, on the same calendar day as the\r\n<br />entry date in the URL, and at most 24 hours after BedTime.",
            "format": "date-time"
          },
          "sleepLatencyMinutes": {
            "type": "integer",
            "description": "How long it took to fall asleep, in minutes; from 0 up to the time between bedtime and wake-up.",
            "format": "int32"
          },
          "awakenings": {
            "type": "integer",
            "description": "Number of times the user woke up during the night, from 0 to 50.",
            "format": "int32"
          },
          "quality": {
            "type": "integer",
            "description": "Subjective sleep quality, an integer from 1 (worst) to 5 (best).",
            "format": "int32"
          },
          "notes": {
            "type": "string",
            "description": "Optional note of at most 2000 characters after removing leading and trailing whitespace, and stored that way. `null`, empty or whitespace-only text means\r\n<br />no note; on `PUT` it removes an existing note."
          }
        },
        "description": "Data of one night in the sleep diary, sent when recording (`POST`) or replacing (`PUT`) an entry. The entry date is taken from the URL."
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. knowledge.material.archived; clients branch on it (ADR-0015)."
          },
          "traceId": {
            "type": "string",
            "description": "Trace identifier of the request, used to find its traces and logs."
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT access token issued by the CIAM, passed unchanged by the edge gateway and BFF.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}